Enhance CUDA installation process in ddistro_install.sh
Updated CUDA version handling and installation process for PyTorch and related packages. Added user prompt for CUDA version selection and modified installation commands accordingly.

+++ ddistro_install.sh
@@ -24,26 +24,47 @@ if [[ "$gpu_answer" =~ ^[Yy][Ee][Ss]$ || "$gpu_answer" =~ ^[Yy]$ ]]; then
     pip check || true
     # Ensure fallback audio loader is available
     pip install --no-cache-dir soundfile
-else
-    cu_tag="cu128"
-    torch_url="https://download.pytorch.org/whl/${cu_tag}"
+    sed -i 's/checkpoint = load_fsspec(model_path, map_location=torch.device("cpu"))\["model"\]/checkpoint = load_fsspec(model_path, map_location=torch.device("cpu"), weights_only=False)["model"]/' /home/dwemer/python-tts/lib/python3.11/site-packages/TTS/tts/models/xtts.py
 
-    python3 -m pip install --upgrade pip wheel ninja virtualenv
-    pip install setuptools==68.1.2
-    # Install app requirements without auto-pulling torch/torchaudio from deps
-    pip install --no-deps -r requirements.txt
-    # Pin to stable, CUDA-tagged PyTorch/Torchaudio that do not require TorchCodec
-    pip cache purge || true
-    pip uninstall -y torch torchaudio torchcodec torchvision || true
-    #pip install --index-url "$torch_url" torch torchaudio torchcodec torchvision 
-    pip install torch==2.7.0+cu128 torchaudio==2.7.0+cu128 torchvision==0.22.0+cu128 --index-url=https://download.pytorch.org/whl/
-    pip check || true
-    # Ensure fallback audio loader is available
-    pip install --no-cache-dir soundfile
-    #pip install xtts-api-server #Fails
+else
+    read -p "New: Use CUDA13. Needs windows updated drivers. RTX 50XX should use this. Usey cuda13? (yes/no): " gpu5_answer
+    if [[ "$gpu5_answer" =~ ^[Yy][Ee][Ss]$ || "$gpu5_answer" =~ ^[Yy]$ ]]; then
+	    cu_tag="cu130"
+	    torch_url="https://download.pytorch.org/whl/${cu_tag}"
+	    echo "Using torch: $torch_url"
+	    python3 -m pip install --upgrade pip wheel ninja virtualenv
+	    pip install setuptools==68.1.2
+	    # Install app requirements without auto-pulling torch/torchaudio from deps
+	    #pip install --no-deps -r requirements5.txt  --index-url=$torch_url
+	    pip install -r requirements5.txt  --extra-index-url=https://download.pytorch.org/whl/cu130
+	    # Pin to stable, CUDA-tagged PyTorch/Torchaudio that do not require TorchCodec
+	    pip check || true
+	    # Ensure fallback audio loader is available
+	    pip install --no-cache-dir soundfile
+	    #pip install xtts-api-server #Fails
+
+    else
+	    cu_tag="cu128"
+	    torch_url="https://download.pytorch.org/whl/${cu_tag}"
+	    echo "Using torch: $torch_url"
+	    python3 -m pip install --upgrade pip wheel ninja virtualenv
+	    pip install setuptools==68.1.2
+	    # Install app requirements without auto-pulling torch/torchaudio from deps
+	    pip install --no-deps -r requirements.txt
+	    # Pin to stable, CUDA-tagged PyTorch/Torchaudio that do not require TorchCodec
+	    pip cache purge || true
+	    pip uninstall -y torch torchaudio torchcodec torchvision || true
+	    #pip install --index-url "$torch_url" torch torchaudio torchcodec torchvision 
+	    pip install torch==2.7.0+cu128 torchaudio==2.7.0+cu128 torchvision==0.22.0+cu128 --index-url=https://download.pytorch.org/whl/
+	    pip check || true
+	    # Ensure fallback audio loader is available
+	    pip install --no-cache-dir soundfile
+	    #pip install xtts-api-server #Fails
+	    sed -i 's/checkpoint = load_fsspec(model_path, map_location=torch.device("cpu"))\["model"\]/checkpoint = load_fsspec(model_path, map_location=torch.device("cpu"), weights_only=False)["model"]/' /home/dwemer/python-tts/lib/python3.11/site-packages/TTS/tts/models/xtts.py
+
+    fi
 fi
 
-sed -i 's/checkpoint = load_fsspec(model_path, map_location=torch.device("cpu"))\["model"\]/checkpoint = load_fsspec(model_path, map_location=torch.device("cpu"), weights_only=False)["model"]/' /home/dwemer/python-tts/lib/python3.11/site-packages/TTS/tts/models/xtts.py
 
 cp /home/dwemer/TheNarrator.wav speakers/TheNarrator.wav
 
@@ -58,6 +79,8 @@ echo "Then close this window. Press ENTER to continue"
 read
 
 echo "please wait...."
+
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/home/dwemer/python-tts/lib/python3.11/site-packages/nvidia/cu13/lib/
 
 python -m xtts_api_server --listen
